requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.1
chromadb==0.4.18
sentence-transformers==2.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.1.0
langchain-text-splitters==0.0.1
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==0.24.0
httpx==0.25.2